Remove swizzles from lowering pipeline
Per #957, swizzles are bad for performance on MI-200, especially when
dealing with GEMMs (which don't have a lot of address calculations to
hide the data movement in), and never really gave us much to begin
with.

They were probably OK on MI-100 because of the agpr->vgpr movements.

They also make delaying vectorization analysis until after fusion
complicated.

So, let's delete them from the lowering pipeline.

We are keeping the operations themselves around because they're good
examples and because we might find a use for them later.