Add ConstDim operator
This operator takes no inputs (no dimensions in the view) and adds
constant-valued dimensions into the output (underlying tensor/next
view).

The operator is parameterized by {constant_value, dimension_length,
constant_value, dimension_length, ...}.

The point of all this is to allow us to nicely represent certain maps,
like "() -> (0, 0, 0, 0)", that you find in various fusion scenarios.

A commit that uses this may or may not be upcoming.