[mlir][Tosa] Split tosa.reshape conversions into tosa-to-tensor
The conversion of tosa.reshape into a combination of
tensor.collapse_shape and tensor.expand_shape operations is currently
handled by tosa-to-linalg. However, this conversion has nothing to do
with the linalg dialect and is instead a conversion to Tensor dialect
operations. In addition, tosa to linalg conversions generate
tosa.reshape operations, which it then converts.

With this change, we move the conversion to the tosa-to-tensor pass,
which can be run after tosa-to-linalg.

This also has the advantage that downstream passes that need to expand
reshape operations to collapse and expand shape but not run all of
tosa-to-linalg.