Update global load/LDS store thread grouping to improve coalescing
Per @giuseros, grouping threads of a workgroup into a D x K group
unconditionally causes failures to coalesce when vectorizing in the D
dimension. Further consideration revealed that this grouping is not
required: the only requirement is that the threads be grouped in the
same way for global load and LDS store.

This commit fixes this unnecessary specificity.

A future followup will update the tiebreak logic to account for what
GEMM dimension has actual unit stride.