Adds tosa reduce_sum to rock reduce conversion (#956)
This commit adds support for converting tosa.reduce_sum
to rock reduce.

Moreover, a new rock attribute (maybe temporary) to indicate
args of kernels where it requires to be initialized to some
value prior to launch. It is verified by doing the same in
the clone verification flow by adding linalg.fills.