Use COMgr to compile our GPU binaries
Since we're considering shipping as a ROCm package, I've investigated
if it's possible to use COMgr instead of shipping our own copy of
LLVM.

This branch shows that it's possible, but I'm not convinced we want to
merge it.

The good news is that the setup works.

The bad news is that, because our LLVM copy is *newer* than the one
shipped in ROCm 5.4 (which is what we're using for our CI) we'll
sometimes need to **downgrade** the LLVM IR we produce to be
acceptable input to ROCm's LLVM. If theirs were newer, it'd be fine,
because LLVM has strong backwards compatibility guarantees, but this
is the other direction.

On top of that, I had to do a few ugly hacks to make COMgr do the
compiler invocations we want.

So ... yeah.