[mlir][Memref] Add memref.reinterpret_elements
memref.reinterpret_elements is a safe interface for certain types of
type punning that rely on either viewing memory as a different type
that has the same bitwidth or on "splitting" larger types into smaller
ones.

For example, a memref<... x vector<4xf32>> could be viewed as a
memref<... x 4 x f32> or a memref<... x 2 x vector<2xf32>> without
incurring any padding concerns, since a 4 x f32 will - if such a type
exists on a target architecture - be stored without padding. However,
a vector<3 x f32> may have implicit padding, so attempting to view it
as its scalar components is forbidden by default. Similarly, turning a
memref<... x f32> into a memref<... x i32> is safe, as operations on
both those memrefs will produce identical memory accesses.

This operation is introduced to allow optimal handling of workgroup
buffers on (AMD) GPUs, where allocating, for example, a
memref<128 x vector<4 x f32>> allows the compiler to infer alignment
requirements in a way that a memref<16384 x i8> (which memref.view
would require) does not. With such an allocation, parts of the code
may need to store to sub-elements of these vectors, even though
vectors will be read as units.

The alignment requirements also cause problems for
vector.transfer_read/transfer_write, which cannot know that the buffer
we are using will be suitably aligned for vector access without this
mechanism.

The main danger of this operation is that it could permit unaligned
loads on archetuctures that do not allow reading the individual bytes
in a word.