Allocates LDS buffers as bytes and view them as appropriate vectors
In order to generate more vectorized reads from LDS (using things like
ds_read2_b64), we need LLVM to be using simpler address computation
logic, namely, we need the getelementpointer operations we're doing to
be vector-typed. This doesn't necessarily mean we need to allocate our
LDS buffers as vector-typed values, but it does mean that we need to
do the view conversion instead of calling vector.transfer_read /
vector.load , as this leads to computing a scalar index and then
loading a vector based off of it.

Since we don't need to actually make the buffer vector-typed, we can
instead use a buffer of bytes and the existing `memref.view` op, which
is intended for exactly this sort of situation (and may have been
motivated by similar LDS-using code in someone else's kernel).

Since our kernels have a bunch of nice powers of 2 running around, we
don't necessarily need to explicitly pass the alignment of the global
to LLVM, but I'm fixing to do that in a followup commit anyway.

The LDS matrix wrapping and LDS store loop wrapping functionality has
been merged into one larger function to make this work - partly
because that's how the previous PR did it.

There's also a minor extension to the vectorizer to handle the fact
that we have a scalar iteration map but vector-valued buffers.

Superscedes PR #1037