Adds support for tranpose+reshape folding
Currently the transposes on inputs are only folded
if they the transpose output immediatedly reaches
a matmul or a conv2d. If they encounter, a reshape
in the middle the folding stops. This will confuse
the gemm-like op.

This commit adds further folding to fold reshapes
i.e. expand/collapse shape ops in between.