Adds support for transpose+reshape folding (#1042)
* Adds support for tranpose+reshape folding

Currently the transposes on inputs are only folded
if they the transpose output immediatedly reaches
a matmul or a conv2d. If they encounter, a reshape
in the middle the folding stops. This will confuse
the gemm-like op.

This commit adds further folding to fold reshapes
i.e. expand/collapse shape ops in between.

* * dropped expand shape support for now
* added conv2d tests.

* * address review feedbacks

* * generalized transpose dim propogation thru collapses.
* added a test case where lower order dims are collapsed.