Add API for quick tuning, otherwise improve the MIGraphX API
Note that this PR contains breaking changes (though, generally, minor
ones) to MIGraphX, and the version define in the headers has been
updated accordingly.

To summarize:
1. The tuning API has been split so that the full and quick tuning
spaces are now separate objects with their own indexing, instead of
having quick be a subset of full (which assumed that we'd have a
sorting metric for the tuning space, which we don't).
Therefore, mlirRockTuningParamGet has been renamed to
mlirRockTuningParamGetFull and there is a mlirRockTuningParamGetQuick
2. The quick tuning space exists and is populated using our
heuristic-mode logic, which uses a hard-coded list of configs per type
and sorts them to prioritize kernels that don't require padding.
3. Both tuning spaces now filter out performance configurations that
cannot result in valid kernels, but do **not** filter out every such
config (they only remove those whose invalidity is clear from the
parameters)
4. A new function has been added to construct the applicability
pipeline, which is used to detect invalid performance configurations.
When tuning, this should be run after applying a performance
configuration to determine if that configuration is valid.
5. Various `int` parameters have been updated to more meaningful types.

Internal details for us:
1. The various PopulateParams classes have grown an
isParamArtrPlausible() method that runs the AffixTuningParameters
checks on a given perf config. This allows us to filter the tuning
space.
2. The "prioritize paddilng-less kernels" sorter has now been made a
public API so the tuning space constructrors can use it.
3. The InitParams* -> tuning parameters attribute constructor is now
exposed for the non-xdlops case as well.
4. The tuning loop now iterates over these InitParams* structs, which
are direct reflections of the perf config string.
5. Some of the C API tests have been updated, expaded, and, in one
case, fixed.