Make packing of LDS store reg buffer accept a view
Currently, the function packLoadBufferToStoreBuffer takes
an assumption of the layout of the load buffer.
However, for fused attention, the load buffer for second
gemm needs to have a flexible layout.
Therefore this commit changes the code, removing the assumption
and make the packing accept a view of the load buffer.