API reworks per Paul
1. The tuning database doesn't need to exist for looking up the
problem string anymore, because that was silly caching that wasn't
needed.
2. String-returning functions take a char* and a length of
caller-owned memory to copy the result in to, and return a size_t of
how many bytes they'd like to have copied.
2b. Constants for how big we think those buffers should be are defined.
3. Internally, a bunch of stuff now takes pointers to
maybe-stack-allocated buffers.
4. A bunch of uses of std::string (aka pointless heap allocations)
have been migrated to their LLVM equivalents. This also saves on
useless copies.
5. The problem string creator now returns LogicalResult() like we'd
want it to.