[DO NOT SQUASH] Static library build improvements & spring cleaning (#1141)
* Exclude our tools from the default static binary build

Though the generous application of EXCLUDE_FROM_ALL when
BUILD_FAT_LIBROCKCOMPILER is on (enabled by the new add_rocmlir_tool
function) we can exclude tools like rocmlir-opt from the build in
static library mode while keeping them around for the dynamic case,
thus removing a bunch of dialects from the build (though some, like
ARM, get pulled in because of conversions like VectorToLLVM).

Overall, this saved about 200 targets from the build.

Next steps including getting the RocMLIROptMain library out of the
build and turning off convolution wrappers.

We might also be able to turn off the X86 target, but that's probably
a separate PR.

* Disable conv validation wrappers, opt main in static lib build

Turn off building and installing libconv-validation-wrappers.so and
libRockMLIROptMain.a when doing a fat binary build. They can still be
built if you ask for them - for example, if you call for tests.

* [mlir] Don't make the ROCm conversions depend on the execution engine

During a conversion to MLIR_ENABLE_EXECUTION_ENGINE from checking for
the native target, the ROCm conversion passes (--serialize-to-hsaco)
were mistakenly flagged for being disabled if the execution ending is
not being built.

These passes use LLVM to build binaries for AMD GPUs, and so require
that backend to be enabled. However, they do not produce native code,
nor do they interact with the JIT or any of the execution engine
support libraries.

When building MLIR into a compiler library that's intended to produce
GPU binaries, we want to build only the AMDGPU backend and have the
binary serialization passes available. This change makes that
possible.

It looks like the CUDA path might currently require a native target,
it's hard to tell, so this commit leaves that if statement untouched.

* Disable the X86 backend when building the static library

The static library isn't using the JIT, and so doesn't need the X86
target, so we save on build space by turning it off.

* Fix tests for the static lib case.

Allow `ninja check-rocmilr` to work in a static library build where we
don't have `mlir-cpu-runner`.

Remove the old `ROCMLIR_DRIVER_ENABLED` setting which is extremely
archaic and wasn't doing anything.

Remove a dependency on mlir-translate that we haven't needed in a good
long time (we use rocmlir-translate now instead).

Add sanity-checking to make sure we don't try to enable E2E tests in
an incoherent way.

Get rid of some get_property() lines that weren't being used

Add enable_rock_driver_any_e2e_test to allow tests to run both on PR
and nightly testing.

Fix lowering_affix_parameters.mlir to take a fixed architecture so it
can be run on the static library codebase.

* Add running static library tests to the CI

This'll give us a smoke test to make sure we don't accidentally break
the static library build.

It'll also ensure we don't accidentally add E2E tests that aren't
properly guarded by the E2E enables.

* Slim down the dependencies, removing stray \*ToLLVMTranslations that we don't use

* Fix the CI I added

* Since we're not doing MIOpen, don't restrict stacic library build to gfx906