[mlir][Arith] Change F32 to BF16 truncation to match __truncsfbf2
The current implementation of truncf %* : f32 to bf16 in the expand
ops pass is slow and does not match the __truncsfbf2 / float2bf16
function in the execution utilities. This has caused divergences in
how floats are truncated between code that uses this pass and code
which relies on __truncsfbf2 implementations, creating spurious
failures in conformance testing.