Fix sorting of Embed{} inputs in vectorization
Fixes https://github.com/ROCmSoftwarePlatform/rocMLIR-internal/issues/1049

The issue found in that bug related to how the dimensions that were
the arguments to Embed{} were sorted for processing. The old logic
used the position in the embed arguments as the tiebreaker for
arguments with equal coefficients, which caused missed vectorization
opportunities as shown in the above bug.

This commit adds improved sorting logic to prevent this sort of thing.