Erase dead transforms during rock-regularize
Fixes
https://github.com/ROCmSoftwarePlatform/rocMLIR-internal/issues/940
and is highly likely to resolve
https://github.com/ROCmSoftwarePlatform/rocMLIR-internal/issues/1072 (but
that'll require testing).

The crashes were tracable to branching read chains (a transform op
with multiple users), which are an issue that we punted on. However,
we did not need to resolve this issue in this commit to fix the bug.

The underlying issues was that there were dead transform ops (the
originals of the operations we inverted) cluttering up use lists and
causing the "a buffer is read multiple times with different transform
chains, we never wrote that code" detector to fire spuriously. This
commit explicitly erases transform ops post-inversion in order prevent
this crash.

The actual assertion failure came from the fact that the
multiple-readers detector would simply cut off the chain collection
early, and then the assert would observe that we never found an actual
reader in at least one of the chains. It wasn't the most meaningful
assert, but it did do its job at spotting an issue.

We also witch up the rock-regularize pattern to collect the inverses
before creating any new IR, in the interests of debugability and the
fact that upstream MLIR would really prefer we not modify IR in
failing patterns (perhaps we should enable those checks in debug
builds...).

We also add the original bug to our nightly testing.