Handle linalg.generic blocks that reduced to constant initilizers
Fixes
https://github.com/ROCmSoftwarePlatform/rocMLIR-internal/issues/1098

Some linalg.generics blocks can (due to, as far as I can tell,
weirdness involved in how broadcasting, linalg.generic fusion. and
such are handled) turn into constant initializers.

The existence of such blocks broke our assumption that all the niputs
to a linalg.generic either traced to the threadwise_write_all from a
gemm or to a function input. This failed to account for the cases
where the input was a local allocation that was filled with constants
by a linalg.generic. (It also failed to account for the possibility of
the value coming from a `memref.get_global`, which might happen if we
end up with large tensor constants in our models).

This commit resolves the issues as follows:
1. Adding special logic to the handling of linalg.generic operations
in -rock-align tiling. When such an operation has no inputs, we
replace the memref.alloc it writes to (which will be the immediate
argument to that op, thanks to regularization) with a rock.alloc in
registers of the same type but take no other action.
2. Improve moveTransformsBefore() to handle the said allocations. If a
transform chain terminates at a private allocation, we move it before
the corresponding threadwise_write_all if it isn't already there, and
move the initializing linalg.generic along with it. (As noted in the
comments, there's a hypothetical failure mode here involving
non-constant initialization, but I'm taking that risk for now).

A minimized version of the crash reproducer has been added as a pass test.