Fix handling of linalg.generic blocks not tied to gemm inputs
This commit fixes the handling of linalg.generic blocks where none of
the inputs participate in a chain of writers that lead to a gemm, but
that feed into such chains. For example, the linalg.generic blocks
that ultimately result from the `migraphx.add(%arg2, %arg3)` in the
example test.

The general solution requires us to add a second case to the
linalg.generic fusion matching. The fusion logic would previously scan
the inputs of a linalg.generic for a threadwise_write_all, which it
would use to get a tile size, set of global coordinates, and a set of
transformations that had to be applied to those arguments in order to
make them all indexable with said global coordinates.

Now, we support a second case: tracing from the output of a
linalg.generic to a threadwise_read_into that was generated by another
round of fusion. That threadwise_read_into then gives us the tiling
information we need to efficiently load a relevant tile of any
additional constant arguments.

Implementing this "visit a generic that we can't do
anything with but might handle later" pattern required control over
the pattern matching worklist that none of the upstream pattern
rewrites could provide. Therefore, this commit defines its own
subclass PatternRewriter - the LinalgAlignRewriter, that lets us
schedule revisits after an initial top-down pass. This rewriter also
provides failure behavior that is more in keeping with how we're using
the rewrite patterns - failure of any rewrite indicates malformed IR
that should fail the fusion pass, as opposed to the current approach
where it leads to stray memref.alloc() ops that throw errors near the
backend.

The fusion flow is relatively unchanged, aside from some variables
being remaned for clarity. The main changes are that the code which
creates threadwise_read_into operations now goes in and schedules the
linalg.generic that wrote the buffer (if any) for re-processing.

Other changes including traceToThreadwiseWriteAll() to
traceToWriter(), removing a function that was equivalant to
untransform().

The fusion flow is also changed to move the writes after the relevant
linalg.generic instead of trying to clone things before the write so
as to prevent use-before-def problems that may have arisen with the
old approach.

We also fix handling of 0-D inputs to threadwise_read_into and
threadwise_write_all in blockwise-gemm-to-threadwise since I needed
that in a previous iteration of the PR and there's no reason to let
that code go to waste.

There is also now a new version of untransform() that doesn't package
the attributes into an ArrayAttr.

Fixes #1188

Fixes
https://github.com/ROCmSoftwarePlatform/rocMLIR-internal/issues/1098

Typo fixes

Stash for merge

Tests fixed

e2e test