Reduce LDS bank conflicts (#1209)
* Reduce LDS bank conflicts

* Addressing review feedbacks

* Addressing review feedbacks - 2

* Addressing review feedbacks - 3

* Addressing review feedbacks - 4

* Fix performance bug in vectorization.

When we rotate/swap dimensions we might put a wedge into the vectorization
process. E.g., if if maxVecLen=16 and kpack=8, each thread is writing 2
contiguous vectors (each of 4 elements) in LDS. LLVM can vectorize this
by a single ds_write_128. However, if we swap thread-id and iter-id dimensions
we will loose this vectorization opportunity.