Improve our annotations of memory usage
A lot of our existing annotation scheme for memory accesses was
scattered all over the place. This commit unifies things into
-rock-analyze-memory-use.

This pass covers:
1. Walking through global_load and global_store operations in order to
find all the "needs 64 bit index" attributes and moving that
declaration to the function level. That means we won't have issues
with the fact that that runs in blockwise-gemm-to-threadwise.
2. Detecting readonly and writeonly arguments. This is done by looking
for arguments that are only (ignoring views) arguments to global_load
or global_store(set) respectively, and then putting those annotations
onto the function arguments for lowering to LLVM.
3. Setting all sorts of relevant LLVM annotations (noalias, nonnull,
noundef, dereferencable, and so on) in order to promise the compiler
things we know are true anyway (if they're not, that's on the client).

Then, this commit also expands our LLVM preparation pass.
The pass is made to run on llvm.func ops, and we use the nesting
mechanism to ensure that we run that on functions inside GPU modules
specifically.

The preparation pass is updated to
1. Update the alignment of vector loads and stores to be the alignment
of the vector, not its component, since MLIR's lowering is
pessimistic in a way we don't want to be.
2. Set !invariant.load on loads from readonly values to given out even
stronger hints about the ability to hoist or sink operations.
3. Work around a shortcoming in how the AMDGPU backend handles
`noalias` (specifically, it gets dropped on kernel arguments during
target-specific lowering) by adding our own alias scopes, which
indicate that loads from one pointer or buffer can't alias the loads
from the other ones. This will hopefully reduce the number of
scheduling DAG edges and therefore give us some of our compile
performance back ... or at least it'll give us better codegen thanks
to more accurate alias information.

This commit also adds the "nontemporal" attribute to rock.global_store
so that -rock-sugar-to-loops can set it on the underlying stores.