Fix returned lengths of the tuning string APIs
Fixes https://github.com/ROCmSoftwarePlatform/rocMLIR-internal/issues/1129

The documentation said we wouldn't add the trailing null byte to the
size of the string returned, but the implementation did do this.

This commit should fix the issue.

It might also fix the lack of performance gain from tuning, if the
null byte was sneaking in.