Loosen the definition of migraphx.dequantizelinear, liberalize lowering
Fixes
https://github.com/ROCmSoftwarePlatform/rocMLIR-internal/issues/1143

Per that ticket, MIGraphX is planning to change its dequantize operator to
not require the input and bias to have the same type. This commit makes
that change in our compilation process so what we won't break when they
make this change.

This will have no impact in the current state of the MIGraphX repository.