Embed device libraries into build to minimize runtime requirements
Fixes
https://github.com/ROCmSoftwarePlatform/rocMLIR-internal/issues/1142

Patch SerializeToHsaco (we'll need to do a version of this again when
we move to the new compilation infrastructure) and its cmake to
find the device libraries at bulid time (if possible) and embed
them (by way of a convenient little Python script) into the
compilation of SerilazieToHsaco. This'll allow our stuff to work even
when there's no /opt/rocm or we can't find it.