Fix attention padding kernels
Padded attention kernels wont just
work out of the box as (implicit) gemm
kernels.

This is because attention kernels does a
softmax normalization after the first gemm
and having additional zeros -- zero not being
the least value in the range -- is going to affect
the normalization.

Therefore, this commit introduces a mid-kernel
overwrite to first gemm output to make the out
of bound values to be negative infinity.