Fix handling of mid-tensor unit dims in transpose/collapse folding
This commit fixes one of the failure cases in
https://github.com/ROCmSoftwarePlatform/AMDMIGraphX/issues/2365 (specifically,
it handles the crash, not the numerical accuracy problems, which are
still under investigation)

The cause of the crash was attempts to transpose merging on the
following sequence of operations:

%t = transpose(%rawA) [0, 1, 3, 2] : tensor<1x1xKxM> to tensor<1x1xMxK>
%a = collapse_shape %t [[0, 1], [2], [3]] : tensor<1x1xMxK> to
rock.gemm %c = %a * %b ...

The problem arose because the logic in the collapse shape folder
didn't correctly handle the case where the unit dimensions it peeled
out of the collapse.

The peeling code code was probably needed so that we could still, for
example, combine transpose [0, 2, 3, 1] : tensor<Ax1xBxC> to
tensor<AxBxCx1> with collapse_shape [[0, 1], [2], [3]], even though a
naive remapping of the reassociation experessions would lead to
non-continuous dimensions being merged.
The problem was that it created the set of reassociation expressions
[[0], [2], [3]], which triggered assertions in tensor::CollapseShape
that all the dimensions (other than the ones on the ends) had to be
listed in order.

This should fix the issue in general, but it's possible we'll want to
investegate all that code and rip it out if we do problem key v2.

(Also, update a utility script to make sure that static library builds
actually rebuild the static library)