[Attention] Persist Q tile in regs if KPerBlock == head_dim (#1311)
The focus of the PR is that when the KPerBlock is equal to head_dim (K dim of the first gemm),
to prefetch the Q tile ( the tile of A tensor of first gemm) directly to regs outside the main (outermost)
loop of the attention kernel. This will remove the need to do repeatitve loading of Q in the main loop
of the attention kernel.

However, the current semantics of blockwise_gemm requires the operands to be in the LDS. Therefore,
this change required to break the first blockwise_gemm into its lowered version and lift the loading of
the first operand (LDS->regs) out of the loop.